okhttp-spring-boot-autoconfigure/src/main/java/org/springframework/boot/autoconfigure/retrofit/OkHttpAutoConfiguration.java
/*
 * Copyright 2016 the original author or authors.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package org.springframework.boot.autoconfigure.retrofit;

import okhttp3.OkHttpClient;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
import org.springframework.boot.context.properties.EnableConfigurationProperties;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;

import javax.net.ssl.SSLContext;
import java.util.concurrent.TimeUnit;

/**
 * {@link EnableAutoConfiguration Auto-configuration} for OkHttp.
 *
 * @author troinine
 */
@Configuration
@ConditionalOnClass(OkHttpClient.class)
@EnableConfigurationProperties(OkHttpProperties.class)
public class OkHttpAutoConfiguration {
    @Autowired(required = false)
    private SSLContext sslContext;

    @Autowired
    private OkHttpProperties okHttpProperties;

    @Bean
    @ConditionalOnMissingBean
    public OkHttpClient okHttpClient() {
        OkHttpClient.Builder builder = new OkHttpClient.Builder()
                .connectTimeout(okHttpProperties.getConnectionTimeout(), TimeUnit.MILLISECONDS)
                .readTimeout(okHttpProperties.getReadTimeout(), TimeUnit.MILLISECONDS)
                .writeTimeout(okHttpProperties.getWriteTimeout(), TimeUnit.MILLISECONDS);

        if (sslContext != null) {
            builder.sslSocketFactory(sslContext.getSocketFactory());
        }

        return builder.build();
    }
}
